+++ engine/enumerate/nenumconstraint.h
@@ -88,7 +88,7 @@ namespace regina {
  *
  * \ifacespython Not present.
  */
-class REGINA_API NEnumConstraintList : public std::vector<std::set<unsigned> > {
+class REGINA_API NEnumConstraintList : public std::vector<std::set<size_t> > {
     public:
         /**
          * Creates an empty list of constraints.
@@ -101,7 +101,7 @@ class REGINA_API NEnumConstraintList : public std::vector<std::set<unsigned> > {
          *
          * @param size the number of constraints to include in the new list.
          */
-        NEnumConstraintList(unsigned size);
+        NEnumConstraintList(size_t size);
 };
 
 /*@}*/
@@ -111,8 +111,8 @@ class REGINA_API NEnumConstraintList : public std::vector<std::set<unsigned> > {
 inline NEnumConstraintList::NEnumConstraintList() {
 }
 
-inline NEnumConstraintList::NEnumConstraintList(unsigned size) :
-        std::vector<std::set<unsigned> >(size) {
+inline NEnumConstraintList::NEnumConstraintList(size_t size) :
+        std::vector<std::set<size_t> >(size) {
 }
 
 } // namespace regina
